+++ quizwithmoxy/src/main/java/harelchuk/maxim/quizwithmoxy/view/SetBankView.java
@@ -6,6 +6,6 @@ import com.arellomobile.mvp.viewstate.strategy.StateStrategyType;
 
 public interface SetBankView extends MvpView {
     @StateStrategyType(AddToEndSingleStrategy.class)
-    void showDebit(boolean is_debit, int time_to_increase,
-                   int debit_GD, int debit_AD, int debit_CP, int user_money);
+    void showDebit(boolean is_debit, long time_to_increase,
+                   long debit_GD, long debit_AD, long debit_CP, long user_money);
 }
